Refactor utils.protocols to typing.protocol
