+++ src/ikigai/typing/__init__.py
@@ -2,3 +2,6 @@
 #
 # SPDX-License-Identifier: MIT
 
+from ikigai.typing import api, protocol
+
+__all__ = ["api", "protocol"]
