+++ src/ikigai/typing/protocol.py
